# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: Meatway API
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/v1/specialmeats/{id}:
    get:
      tags:
      - specialmeat-controller
      summary: Вывести информацию объявления cо специальным мясом
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpecialmeatAdResponse"
    post:
      tags:
      - specialmeat-controller
      summary: Создать объявление cо специальным мясом
      operationId: createById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialmeatAdSaveRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpecialmeatAdResponse"
    delete:
      tags:
      - specialmeat-controller
      summary: Удалить объявление cо специальным мясом
      operationId: deleteById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - specialmeat-controller
      summary: Редактировать объявление cо специальным мясом
      operationId: editById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialmeatAdSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpecialmeatAdResponse"
  /api/v1/sheepmeats/{id}:
    get:
      tags:
      - sheepmeat-controller
      summary: Вывести информацию объявления c бараниной
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SheepmeatAdResponse"
    post:
      tags:
      - sheepmeat-controller
      summary: Создать объявление c бараниной
      operationId: createById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheepmeatAdSaveRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SheepmeatAdResponse"
    delete:
      tags:
      - sheepmeat-controller
      summary: Удалить объявление c бараниной
      operationId: deleteById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - sheepmeat-controller
      summary: Редактировать объявление c бараниной
      operationId: editById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheepmeatAdSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SheepmeatAdResponse"
  /api/v1/porks/{id}:
    get:
      tags:
      - pork-controller
      summary: Вывести информацию объявления cо свининой
      operationId: findById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PorkAdResponse"
    post:
      tags:
      - pork-controller
      summary: Создать объявление cо свининой
      operationId: createById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PorkAdSaveRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PorkAdResponse"
    delete:
      tags:
      - pork-controller
      summary: Удалить объявление cо свининой
      operationId: deleteById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - pork-controller
      summary: Редактировать объявление cо свининой
      operationId: editById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PorkAdSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PorkAdResponse"
...
  /api/v1/birds/{id}:
    get:
      tags:
      - bird-controller
      summary: Вывести информацию объявления c птицей
      operationId: findById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BirdAdResponse"
    post:
      tags:
      - bird-controller
      summary: Создать объявление c птицей
      operationId: createById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BirdAdSaveRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BirdAdResponse"
    delete:
      tags:
      - bird-controller
      summary: Удалить объявление c птицей
      operationId: deleteById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - bird-controller
      summary: Редактировать объявление c птицей
      operationId: editById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BirdAdSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BirdAdResponse"
  /api/v1/beefs/{id}:
    get:
      tags:
      - beef-controller
      summary: Вывести информацию объявления c говядиной
      operationId: findById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BeefAdResponse"
    post:
      tags:
      - beef-controller
      summary: Создать объявление c говядиной
      operationId: createById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeefAdSaveRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BeefAdResponse"
    delete:
      tags:
      - beef-controller
      summary: Удалить объявление c говядиной
      operationId: deleteById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
    patch:
      tags:
      - beef-controller
      summary: Редактировать объявление c говядиной
      operationId: editById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BeefAdSaveRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BeefAdResponse"
...
  /api/v1/specialmeats:
    get:
      tags:
      - specialmeat-controller
      summary: Вывести краткую информацию об объявлениях cо специальным мясом
      operationId: findAll
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialmeatAdsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShortAdsResponse"
...
  /api/v1/porks:
    get:
      tags:
      - pork-controller
      summary: Вывести краткую информацию об объявлениях cо свининой
      operationId: findAll_2
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PorkAdsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShortAdsResponse"
...
components:
  schemas:
...
    SpecialmeatAdSaveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        hasMedicalCertificate:
          type: boolean
    SpecialmeatAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        hasMedicalCertificate:
          type: boolean
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
    SheepmeatAdSaveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        hasMedicalCertificate:
          type: boolean
    SheepmeatAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        hasMedicalCertificate:
          type: boolean
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
    PorkAdSaveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        hasMedicalCertificate:
          type: boolean
    PorkAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        hasMedicalCertificate:
          type: boolean
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
...
    BirdAdSaveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        birdType:
          type: string
        hasMedicalCertificate:
          type: boolean
    BirdAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        birdType:
          type: string
        hasMedicalCertificate:
          type: boolean
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
    BeefAdSaveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        isMramor:
          type: boolean
        hasMedicalCertificate:
          type: boolean
    BeefAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        price:
          type: number
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        location:
          type: string
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        isMramor:
          type: boolean
        hasMedicalCertificate:
          type: boolean
        isActive:
          type: boolean
        creationDate:
          type: string
          format: date-time
...
    SpecialmeatAdsRequest:
      type: object
      properties:
        isRetail:
          type: boolean
        medicalCertificate:
          type: boolean
        isFrozen:
          type: boolean
        priceFrom:
          type: number
          format: double
        priceTo:
          type: number
          format: double
        weightFrom:
          type: number
          format: double
        weightTo:
          type: number
          format: double
        quantityFrom:
          type: integer
          format: int32
        quantityTo:
          type: integer
          format: int32
        monthsAgeFrom:
          type: integer
          format: int32
        monthsAgeTo:
          type: integer
          format: int32
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        isHalal:
          type: boolean
        animalType:
          type: string
    ShortAdResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        price:
          type: number
        animalType:
          type: string
        breed:
          type: string
        monthsAge:
          type: integer
          format: int32
        weight:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
        location:
          type: string
        username:
          type: string
        stars:
          type: number
          format: double
        isFrozen:
          type: boolean
        isRetail:
          type: boolean
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        killDate:
          type: string
          format: date-time
        isHalal:
          type: boolean
        isMramor:
          type: boolean
        isPremium:
          type: boolean
        hasMedicalCertificate:
          type: boolean
        birdType:
          type: string
    ShortAdsResponse:
      type: object
      properties:
        ads:
          type: array
          items:
            $ref: "#/components/schemas/ShortAdResponse"
        size:
          type: integer
          format: int32
...
    PorkAdsRequest:
      type: object
      properties:
        isRetail:
          type: boolean
        medicalCertificate:
          type: boolean
        isFrozen:
          type: boolean
        priceFrom:
          type: number
          format: double
        priceTo:
          type: number
          format: double
        weightFrom:
          type: number
          format: double
        weightTo:
          type: number
          format: double
        quantityFrom:
          type: integer
          format: int32
        quantityTo:
          type: integer
          format: int32
        monthsAgeFrom:
          type: integer
          format: int32
        monthsAgeTo:
          type: integer
          format: int32
        dateBegin:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        fatContent:
          type: string
        processingType:
          type: string
...